WIP: chore: adopt ruff in place of flake8

+++ src/dotlocalslashbin.py
@@ -2,13 +2,9 @@
 # src/dotlocalslashbin.py
 # Copyright 2022 Keith Maxwell
 # SPDX-License-Identifier: MPL-2.0
-"""Download and extract files to ~/.local/bin/"""
+"""Download and extract files to `~/.local/bin/`."""
 import tarfile
-from argparse import (
-    ArgumentDefaultsHelpFormatter as formatter_class,
-    ArgumentParser,
-    Namespace,
-)
+from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser, Namespace
 from collections.abc import Generator
 from contextlib import contextmanager
 from enum import Enum
@@ -27,9 +23,10 @@ from zipfile import ZipFile
 __version__ = "0.0.8"
 
 DEFAULT_OUTPUT = "~/.local/bin/"
+_SHA512_LENGTH = 128
 
 
-class CustomNamespace(Namespace):
+class _CustomNamespace(Namespace):
     output: Path
     input: Path
     downloaded: Path
@@ -38,8 +35,11 @@ class CustomNamespace(Namespace):
 Action = Enum("Action", ["command", "copy", "symlink", "untar", "unzip"])
 
 
-def parse_args():
-    parser = ArgumentParser(prog=Path(__file__).name, formatter_class=formatter_class)
+def _parse_args() -> _CustomNamespace:
+    parser = ArgumentParser(
+        prog=Path(__file__).name,
+        formatter_class=ArgumentDefaultsHelpFormatter,
+    )
     parser.add_argument("--version", action="version", version=__version__)
     help_ = "TOML specification"
     parser.add_argument("--input", default="bin.toml", help=help_, type=Path)
@@ -48,12 +48,12 @@ def parse_args():
     help_ = "Output directory"
     default = "~/.cache/dotlocalslashbin/"
     parser.add_argument("--downloaded", default=default, help=help_, type=Path)
-    return parser.parse_args(namespace=CustomNamespace)
+    return parser.parse_args(namespace=_CustomNamespace())
 
 
 @contextmanager
-def _download(
-    args: type[CustomNamespace],
+def _process(
+    args: _CustomNamespace,
     *,
     name: str,
     url: str,
@@ -63,11 +63,13 @@ def _download(
     version: str | None = None,
     prefix: str | None = None,
     command: str | None = None,
-    ignore: set = set(),
+    ignore: set | None = None,
 ) -> Generator[tuple[Path, Path], None, None]:
-    """Context manager to download and install a program
+    """Context manager to download and install a program.
 
     Arguments:
+        args: the parsed command line arguments
+        name: the name of the target "file"
         url: the URL to download
         action: action to take to install for example copy
         target: the destination
@@ -75,42 +77,47 @@ def _download(
         version: an argument to display the version for example --version
         prefix: to remove when untarring
         command: command to run to install after download
+        ignore: a set of files to ignore when extracting
+
     """
+    if ignore is None:
+        ignore = set()
     if target is None:
         target = args.output.joinpath(name)
-    assert target is not None
     if args.output != DEFAULT_OUTPUT:
         target = target.absolute()
 
     if url.startswith("https://"):
         downloaded = args.downloaded.expanduser() / url.rsplit("/", 1)[1]
-        downloaded.parent.mkdir(parents=True, exist_ok=True)
-        if not downloaded.is_file():
-            with urlopen(url) as fp, downloaded.open("wb") as dp:
-                if "content-length" in fp.headers:
-                    size = int(fp.headers["Content-Length"])
-                else:
-                    size = -1
-
-                print(f"Downloading {name}…")
-                written = dp.write(fp.read())
-
-            if size >= 0 and written != size:
-                raise RuntimeError("Wrong content length")
-
-        if expected:
-            with downloaded.open("rb") as f:
-                digest = file_digest(f, "sha512" if len(expected) == 128 else "sha256")
-
-            if (actual := digest.hexdigest()) != expected:
-                raise RuntimeError(
-                    f"Unexpected digest for {downloaded}: {actual=} {expected=}"
-                )
     else:
         downloaded = Path(url)
 
+    if not downloaded.is_file() and url.startswith("https://"):
+        downloaded.parent.mkdir(parents=True, exist_ok=True)
+        with urlopen(url) as fp, downloaded.open("wb") as dp:
+            if "content-length" in fp.headers:
+                size = int(fp.headers["Content-Length"])
+            else:
+                size = -1
+
+            print(f"Downloading {name}…")
+            written = dp.write(fp.read())
+
+        if size >= 0 and written != size:
+            msg = "Wrong content length"
+            raise RuntimeError(msg)
+
+    if expected:
+        with downloaded.open("rb") as f:
+            _digest = "sha512" if len(expected) == _SHA512_LENGTH else "sha256"
+            digest = file_digest(f, _digest)
+
+        if (actual := digest.hexdigest()) != expected:
+            msg = f"Unexpected digest for {downloaded}: {actual=} {expected=}"
+            raise RuntimeError(msg)
+
     if action is None:
-        if url.endswith(".tar.gz") or url.endswith(".tar"):
+        if url.endswith((".tar.gz", ".tar")):
             action = Action.untar
         elif url.endswith(".zip"):
             action = Action.unzip
@@ -145,7 +152,7 @@ def _download(
                     file.extract(member, path=target.parent)
 
     elif action == Action.command and command is not None:
-        kwargs = dict(target=target, downloaded=downloaded)
+        kwargs = {"target": target, "downloaded": downloaded}
         run(split(command.format(**kwargs)), check=True)
 
     yield downloaded, target
@@ -161,7 +168,8 @@ def _download(
 
 
 def main() -> int:
-    args = parse_args()
+    """Parse command line arguments and download each file."""
+    args = _parse_args()
 
     with args.input.expanduser().open("rb") as file:
         data = load(file)
@@ -173,7 +181,7 @@ def main() -> int:
         if "action" in kwargs:
             kwargs["action"] = getattr(Action, kwargs["action"])
         try:
-            with _download(args, **kwargs) as (downloaded, target):
+            with _process(args, **kwargs) as (downloaded, target):
                 pass
         except HTTPError as e:
             print(f"Error {e.code} downloading {e.url}")
